Refactored ProfileScreenSettings.
