+++ androidApp/src/main/java/com/gorosoft/bookme/now/android/ui/main_screen/profile/LogoutBottomSheet.kt
@@ -1,13 +1,10 @@
 package com.gorosoft.bookme.now.android.ui.main_screen.profile
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.HorizontalDivider
@@ -16,7 +13,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -26,48 +22,31 @@ import com.gorosoft.bookme.now.android.ui.utils.PrimaryButton
 import com.gorosoft.bookme.now.android.ui.utils.SecondaryButton
 
 @Composable
-fun Logout(
+fun LogoutBottomSheet(
     modifier: Modifier = Modifier,
     onLogout: () -> Unit = { },
     onCancel: () -> Unit = { },
 ) {
-
     Column(
         modifier = modifier
             .fillMaxWidth()
-            .wrapContentHeight()
-            .clip(RoundedCornerShape(topStart = 40.dp, topEnd = 40.dp))
-            .background(AppTheme.colors.backgroundThemed.backgroundMain)
-            .border(1.dp, Color(0xFFF5F5F5)),
+            .wrapContentHeight(),
         horizontalAlignment = Alignment.CenterHorizontally,
-
-
-        ) {
-        HorizontalDivider(
-            modifier = Modifier
-                .width(38.dp)
-                .padding(top = 8.dp),
-            color = AppTheme.colors.grayscale.gs200,
-            thickness = 3.dp,
-        )
-
+    ) {
         Text(
-            modifier = Modifier
-                .padding(top = 24.dp),
+            modifier = Modifier,
             text = stringResource(R.string.logout),
             style = AppTheme.typography.heading.h4,
             color = AppTheme.colors.alertStatus.error
-        )
 
+        )
         HorizontalDivider(
             modifier = Modifier
-                .width(380.dp)
+                .fillMaxWidth()
                 .padding(top = 24.dp, start = 24.dp, end = 24.dp),
             color = AppTheme.colors.grayscale.gs200,
             thickness = 1.dp,
         )
-
-
         Text(
             modifier = Modifier
                 .padding(top = 24.dp),
@@ -81,7 +60,6 @@ fun Logout(
                 .clip(RoundedCornerShape(20.dp))
                 .padding(top = 24.dp, start = 24.dp, end = 24.dp, bottom = 48.dp),
             horizontalArrangement = Arrangement.spacedBy(12.dp)
-
 
         ) {
             SecondaryButton(
@@ -102,13 +80,10 @@ fun Logout(
     }
 }
 
-@Preview
+@Preview(showBackground = true, backgroundColor = 0xFFFFFFFF)
 @Composable
-private fun LogoutPreview() {
+private fun LogoutBottomSheetPreview() {
     AppTheme {
-        Logout()
+        LogoutBottomSheet()
     }
 }
-
-
-
