+++ androidApp/src/main/java/com/gorosoft/bookme/now/android/ui/main_screen/profile/ProfileScreen.kt
@@ -29,19 +29,29 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.gorosoft.bookme.now.android.R
 import com.gorosoft.bookme.now.android.annotations.BottomBarNavGraph
+import com.gorosoft.bookme.now.android.ui.destinations.ChooseLanguageScreenDestination
 import com.gorosoft.bookme.now.android.ui.theme.AppTheme
 import com.gorosoft.bookme.now.android.ui.utils.debounceClick
 import com.ramcosta.composedestinations.annotation.Destination
+import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 
 @BottomBarNavGraph
 @Destination
 @Composable
-fun ProfileScreen() {
-    ProfileScreenContent()
+fun ProfileScreen(
+    navigator: DestinationsNavigator,
+) {
+    ProfileScreenContent(
+        navigateToLanguageScreen = {
+            navigator.navigate(ChooseLanguageScreenDestination)
+        }
+    )
 }
 
 @Composable
-fun ProfileScreenContent() {
+private fun ProfileScreenContent(
+    navigateToLanguageScreen: () -> Unit = {},
+) {
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -60,7 +70,7 @@ fun ProfileScreenContent() {
         )
         Spacer(modifier = Modifier.height(24.dp))
         ProfileScreenSettings(
-            onLanguageClick = {},
+            onLanguageClick = navigateToLanguageScreen,
         )
     }
 }
